pyproject.toml
[tool.mypy]
plugins = ["pydantic.mypy"]
mypy_path = "./"
files = ["aiosu"]
disallow_untyped_defs = true
disallow_any_unimported = true
no_implicit_optional = true
check_untyped_defs = true
warn_unused_ignores = true
show_error_codes = true

[tool.isort]
add_imports = ["from __future__ import annotations"]
force_single_line = true
profile = "black"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.poetry]
name = "aiosu"
version = "2.2.0"
description = "Simple and fast osu! API v1 and v2 library"
authors = ["Nice Aesthetics <nice@aesth.dev>"]
license = "GPLv3+"
readme = "README.rst"
repository = "https://github.com/NiceAesth/aiosu"
documentation = "https://aiosu.readthedocs.io/"
keywords = ["osu!", "osu", "api"]
classifiers = [
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
    "Typing :: Typed",
]

[tool.poetry.dependencies]
python = "^3.9"
aiohttp = "^3.8.3"
aiolimiter = "^1.0.0"
emojiflags = "^0.1.1"
orjson = "^3.8.3"
pydantic = "^2.0.3"
pytest = { version = "^7.2.0", optional = true }
pytest-asyncio = { version = "^0.23.0", optional = true }
pytest-mock = { version = "^3.10.0", optional = true }
toml = { version = "^0.10.2", optional = true }
sphinx = { version = "^7.0.0", optional = true }
furo = { version = "^2023.5.20", optional = true }
types-toml = { version = "^0.10.8.1", optional = true }
pyjwt = "^2.6.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
pytest-asyncio = "^0.23.0"
pytest-mock = "^3.10.0"
mypy = "^1.0"
toml = "^0.10.2"
types-toml = "^0.10.8.1"
sphinx = "^7.0.0"
furo = "^2023.5.20"
pytest-cov = "^4.0.0"
black = { version = "^23.0.0", allow-prereleases = true }
pre-commit = "^3.2.2"
python-dotenv = "^1.0.0"

[tool.poetry.extras]
test = ["pytest", "pytest-asyncio", "pytest-mock", "toml", "types-toml"]
docs = ["Sphinx", "furo", "toml"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
